Add (UINT16) static cast before passing UINT32 parameter to CreateDeviceNode
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@11508 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/DevicePathDxe/DevicePathFromText.c
@@ -1158,7 +1158,7 @@ DevPathFromTextAcpiAdr (
   AcpiAdr = (ACPI_ADR_DEVICE_PATH *) CreateDeviceNode (
                                        ACPI_DEVICE_PATH,
                                        ACPI_ADR_DP,
-                                       sizeof (ACPI_ADR_DEVICE_PATH)
+                                       (UINT16) sizeof (ACPI_ADR_DEVICE_PATH)
                                        );
   ASSERT (AcpiAdr != NULL);
